requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.3
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-smartfields==1.1.0
gunicorn==19.9.0
idna==2.8
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.1
urllib3==1.25.3
whitenoise==4.1.3

